pyproject.toml
[project]
name = "frame-ai"
version = "0.1.0"
description = "AI-powered photography coach and analysis tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.3.0",
    "anthropic>=0.40.0",
    "click>=8.0.0",
]
